+++ src/main/kotlin/com/github/texousliu/opengitmoji/config/OpenGMConfig.kt
@@ -6,14 +6,16 @@ import com.github.texousliu.opengitmoji.model.GMInputModel
 import com.github.texousliu.opengitmoji.model.GMLanguage
 import com.intellij.openapi.options.SearchableConfigurable
 import com.intellij.openapi.ui.ComboBox
+import com.intellij.ui.IdeBorderFactory
+import com.intellij.ui.components.fields.ExpandableTextField
 import com.intellij.util.containers.stream
+import com.intellij.util.ui.UI
+import java.awt.BorderLayout
 import java.awt.FlowLayout
 import java.awt.GridLayout
 import javax.swing.JCheckBox
 import javax.swing.JComponent
-import javax.swing.JLabel
 import javax.swing.JPanel
-import kotlin.streams.toList
 
 
 class OpenGMConfig : SearchableConfigurable {
@@ -26,7 +28,8 @@ class OpenGMConfig : SearchableConfigurable {
     private val gmTailTextComboBox = ComboBox(GMField.values().stream().map(GMField::description).toArray())
     private val gmTypeTextComboBox = ComboBox(GMField.values().stream().map(GMField::description).toArray())
 
-    private val gmTriggerConditionBox = JCheckBox("")
+    private val gmTriggerConditionBox = JCheckBox("Trigger start with colon")
+    private val gmSuffixTextField = ExpandableTextField()
 
     private var configGMLanguage: GMLanguage = OpenGMContext.getLanguage()
     private var configGMInputModel: GMInputModel = OpenGMContext.getInputModel()
@@ -35,6 +38,7 @@ class OpenGMConfig : SearchableConfigurable {
     private var configGMTailText: GMField = OpenGMContext.getTailText()
     private var configGMTypeText: GMField = OpenGMContext.getTypeText()
     private var configGMTriggerCondition: Boolean = OpenGMContext.getTriggerCondition()
+    private var configGMSuffixText: String = OpenGMContext.getSuffixText();
 
     override fun isModified(): Boolean = gmLanguageComboBox.selectedIndex != configGMLanguage.ordinal
             || gmInputModelComboBox.selectedIndex != configGMInputModel.ordinal
@@ -42,6 +46,7 @@ class OpenGMConfig : SearchableConfigurable {
             || gmTailTextComboBox.selectedIndex != configGMTailText.ordinal
             || gmTypeTextComboBox.selectedIndex != configGMTypeText.ordinal
             || gmTriggerConditionBox.isSelected != configGMTriggerCondition
+            || gmSuffixTextField.text != configGMSuffixText
 
 
     override fun getDisplayName(): String = "Open Gitmoji"
@@ -49,26 +54,39 @@ class OpenGMConfig : SearchableConfigurable {
     override fun getId(): String = "open.texousliu.config.settings.gm.OpenGMSettings"
 
     init {
-        val flow = GridLayout(20, 2)
-        gmSettingsPanel = JPanel(flow)
+        // input border
+        val input = UI.PanelFactory.grid()
+            .add(UI.PanelFactory.panel(gmTriggerConditionBox)
+                .withComment("Get prompt through text starting with ':' or '：'. Such as ':s'"))
+            .createPanel()
+        input.border = IdeBorderFactory.createTitledBorder("Input")
 
-        val gmTCPanel = JPanel(FlowLayout(FlowLayout.LEADING))
-        gmTCPanel.add(JLabel("Trigger Start With ':' "))
-        gmTCPanel.add(gmTriggerConditionBox, null)
-        gmSettingsPanel.add(gmTCPanel)
-
+        // format border
+        // 输入格式
         val gmModelPanel = JPanel(FlowLayout(FlowLayout.LEADING))
-        gmModelPanel.add(JLabel("Input Model    "))
         gmModelPanel.add(gmInputModelComboBox, null)
         gmModelPanel.add(gmLanguageComboBox, null)
-        gmSettingsPanel.add(gmModelPanel)
-
+        // 提示格式
         val gmTextPanel = JPanel(FlowLayout(FlowLayout.LEADING))
-        gmTextPanel.add(JLabel("Hint Formatter "))
         gmTextPanel.add(gmPresentableTextComboBox, null)
         gmTextPanel.add(gmTailTextComboBox, null)
         gmTextPanel.add(gmTypeTextComboBox, null)
-        gmSettingsPanel.add(gmTextPanel)
+
+        val format = UI.PanelFactory.grid()
+            .add(UI.PanelFactory.panel(gmModelPanel).withLabel("Input format:"))
+            .add(UI.PanelFactory.panel(gmTextPanel).withLabel("Hint format:"))
+            .add(UI.PanelFactory.panel(gmSuffixTextField).withLabel("Suffix append:")
+                .withComment("Add some additional information after entering text. " +
+                        "For example: text, date and time. " +
+                        "The default is a space text. <br/>" +
+                        "Date expression: #{DATE}, " +
+                        "Time expression: #{TIME}"))
+            .createPanel()
+        format.border = IdeBorderFactory.createTitledBorder("Format")
+
+        gmSettingsPanel = JPanel(BorderLayout())
+        gmSettingsPanel.add(input, BorderLayout.NORTH)
+        gmSettingsPanel.add(format, BorderLayout.WEST)
     }
 
     override fun apply() {
@@ -80,6 +98,7 @@ class OpenGMConfig : SearchableConfigurable {
         configGMTypeText = GMField.values()[gmTypeTextComboBox.selectedIndex]
 
         configGMTriggerCondition = gmTriggerConditionBox.isSelected
+        configGMSuffixText = gmSuffixTextField.text
 
         OpenGMContext.setLanguage(configGMLanguage)
         OpenGMContext.setInputModel(configGMInputModel)
@@ -89,6 +108,7 @@ class OpenGMConfig : SearchableConfigurable {
         OpenGMContext.setTypeText(configGMTypeText)
 
         OpenGMContext.setTriggerCondition(configGMTriggerCondition)
+        OpenGMContext.setSuffixText(configGMSuffixText)
     }
 
     override fun reset() {
@@ -100,6 +120,7 @@ class OpenGMConfig : SearchableConfigurable {
         configGMTypeText = OpenGMContext.getTypeText()
 
         configGMTriggerCondition = OpenGMContext.getTriggerCondition()
+        configGMSuffixText = OpenGMContext.getSuffixText()
 
         gmLanguageComboBox.selectedIndex = configGMLanguage.ordinal
         gmInputModelComboBox.selectedIndex = configGMInputModel.ordinal
@@ -109,6 +130,7 @@ class OpenGMConfig : SearchableConfigurable {
         gmTypeTextComboBox.selectedIndex = configGMTypeText.ordinal
 
         gmTriggerConditionBox.isSelected = configGMTriggerCondition
+        gmSuffixTextField.text = configGMSuffixText
     }
 
     override fun createComponent(): JComponent = gmSettingsPanel